achieve publish question in h2 database

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -5,7 +5,7 @@ create table user
 	name varchar(50),
 	token char(36),
 	gmt_create bigint,
-	gmt_modified int,
+	gmt_modified bigint,
 	constraint user_pk
 		primary key (id)
 );
